+++ src/main/java/com/jsongnoti/jsongnoti_web/auth/oauth/PrincipalOauth2UserService.java
@@ -6,6 +6,8 @@ import com.jsongnoti.jsongnoti_web.auth.PrincipalDetails;
 import com.jsongnoti.jsongnoti_web.auth.oauth.provider.GoogleUserInfo;
 import com.jsongnoti.jsongnoti_web.auth.oauth.provider.OAuth2UserInfo;
 import com.jsongnoti.jsongnoti_web.domain.User;
+import com.jsongnoti.jsongnoti_web.domain.enums.Brand;
+import com.jsongnoti.jsongnoti_web.domain.enums.MemoPresentType;
 import com.jsongnoti.jsongnoti_web.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -95,6 +97,8 @@ public class PrincipalOauth2UserService extends DefaultOAuth2UserService {
                     .role("ROLE_USER")
                     .provider(oAuth2UserInfo.getProvider())
                     .providerId(oAuth2UserInfo.getProviderId())
+                    .memoPresentType(MemoPresentType.PRESENT_ORDER)
+                    .showMemoBrand(Brand.TJ) //TODO 나중에 금영 구현하면 ALL 로 수정
                     .build();
 
             userRepository.save(newUser);
