+++ src/main/resources/static/css/header-text.css
@@ -1,7 +1,19 @@
+/* 헤더텍스트 추가 설정 */
+#container nav h1 {
+    padding: 0.5rem 0.5rem;
+    background-color: #fff3cd;
+    border-radius: 10px;
+    /*box-shadow:  0 0 10px 10px #fff3cd;*/
+}
+
+h2, h3 {
+    margin-bottom: 1rem; /* 원본 미디어 쿼리 관계없이 0.5rem */
+}
+
 /* 헤더 텍스트 사이즈 보정 */
 
 h1 .navbar-brand {
-    font-size: calc(1.5rem + 0.6vw);
+    font-size: calc(1.5rem + 0.7vw);
 }
 
 h1 {
@@ -9,7 +21,7 @@ h1 {
 }
 
 h2 { /* 외부헤더 (신곡목록, 사용방법 ...) */
-    font-size: calc(1.4rem + 0.6vw); /* calc(1.325rem + 0.9vw) */
+    font-size: calc(1.425rem + 0.7vw); /* calc(1.325rem + 0.9vw) */
 }
 
 h3 { /* 사용방법-구독하기, ... */
@@ -17,7 +29,7 @@ h3 { /* 사용방법-구독하기, ... */
 }
 
 h4 { /* 모달 헤더, 모달 하위헤더 */
-    font-size: calc(1rem + 0.3vw);
+    font-size: calc(1.05rem + 0.5vw);
 }
 
 .lead {
