+++ src/main/resources/static/js/memo/memo-util.js
@@ -1,3 +1,5 @@
+import SongTableUtil from "../song-table-util.js";
+
 class MemoUtil {
 
     // Table Utils ============================================================
@@ -34,7 +36,7 @@ class MemoUtil {
      */
     refreshMemoTable(brand) {
         let dataList = this.getMemoTableData(brand); // 서버에서 데이터 가져옴
-        let songTable = this.renderEmptySongTable(dataList); // 빈 테이블 복사후 데이터 채움
+        let songTable = SongTableUtil.renderEmptySongTable(dataList); // 빈 테이블 복사후 데이터 채움
 
         // memoId, presentOrder 붙임
         let songRows = songTable.find('tbody tr');
@@ -79,33 +81,6 @@ class MemoUtil {
         return memoData;
     }
 
-    /**
-     * .empty-song-table 을 복사한 뒤 dataList 의 값으로 채움
-     * @param dataList 서버에서 조회한 노래 데이터 리스트
-     * @returns {*|jQuery} 데이터를 채운 테이블
-     */
-    renderEmptySongTable(dataList) {
-        if (dataList == null) {
-            console.log('dataList is null');
-            return;
-        }
-        let emptySongTable = $('.empty-song-table');
-        let resultTable = emptySongTable.clone();
-        resultTable.removeClass('empty-song-table');
-
-        let firstRow = resultTable.find('tbody tr:first-child');
-        $.each(dataList, function (index, song) {
-            let row = firstRow.clone();
-            row.find('.song-number span').text(song.number);
-            row.find('.song-title-text').text(song.title);
-            row.find('.song-info span').text(song.info);
-            row.find('.song-singer').text(song.singer);
-            firstRow.before(row); // after 가 아닌 before 로 붙여야 순서가 맞음
-        });
-        firstRow.remove();
-        return resultTable;
-    }
-
     // Element Utils ==========================================================
 
     toggleSearchForm(command) {
@@ -133,22 +108,6 @@ class MemoUtil {
         } else if (command === 'off') {
             $('.switch-order-button').prop('disabled', true).hide();
         }
-    }
-
-    /**
-     * jquery 객체를 받아서 해당 엘리먼트를 깜빡임
-     */
-    blinkElement(element) {
-        $(element).css({
-            transition: 'background-color 300ms linear',
-            backgroundColor: '#fff3cd'
-        });
-        setTimeout(function () {
-            $(element).css('background-color', '')
-        }, 250);
-        setTimeout(function () {
-            $(element).css('transition', '')
-        }, 500);
     }
 
 }
