+++ src/main/resources/templates/privacyPolicy.html
@@ -17,7 +17,7 @@
   <link href="https://hangeul.pstatic.net/hangeul_static/css/nanum-square-round.css" rel="stylesheet"/>
   <link href="/static/css/styles.css" th:href="@{/css/styles.css}" rel="stylesheet"/>
   <link href="/static/css/containers.css" th:href="@{/css/containers.css}" rel="stylesheet"/>
-  <link href="/static/css/header-text-size.css" th:href="@{/css/header-text-size.css}" rel="stylesheet"/>
+  <link href="/css/header-text.css" th:href="@{/css/header-text.css}" rel="stylesheet"/>
   <link href="/static/css/song-table.css" th:href="@{/css/song-table.css}" rel="stylesheet"/>
   <link href="/static/css/modal.css" th:href="@{/css/modal.css}" rel="stylesheet"/>
   <link href="/static/css/footer.css" th:href="@{/css/footer.css}" rel="stylesheet"/>
